refactor: split game_manager.rs into functional modules
Break the single 4994-line file into server/game_manager/{mod,
scenario_flow, turn_flow, attack_flow}.rs, grouped by concern
(scenario progression, turn/round flow, attack resolution). Each
module carries its own unit tests. eval_end_of_round and
process_no_atk_launched widened from private to pub(super) so
attack_flow.rs can call them. No behavior change — same public
API, same 348 passing tests.
