+++ src/server/game_manager/attack_flow.rs
@@ -1,326 +1,26 @@
-use std::{collections::HashMap, path::Path};
-
 use crate::{
     character_mod::{
         attack_type::{AttackType, LauncherAtkInfo},
         buffers::BufKinds,
-        character::{Character, CharacterKind},
-        class::Class,
-        equipment::{Equipment, EquipmentJsonKey},
-        experience::{build_exp_to_next_level, build_experience},
-        loot::LootType,
         rounds_information::AmountType,
     },
     common::{
-        constants::{character_const::ULTIMATE_LEVEL, paths_const::*, stats_const::*},
+        constants::stats_const::*,
         log_data::{
             LogData,
             const_colors::{DARK_RED, LIGHT_BLUE, LIGHT_GREEN},
         },
     },
     server::{
-        end_of_scenario::{EndOfScenario, LevelUp},
-        game_paths::GamePaths,
-        game_state::{GameState, GameStatus},
-        players_manager::{DodgeInfo, GameAtkEffect, PlayerManager},
-        scenario::{Scenario, ScenarioState},
+        game_state::GameStatus,
+        players_manager::{DodgeInfo, GameAtkEffect},
     },
-    shop::build_consumable_by_name,
     utils,
 };
-use anyhow::Result;
-use serde::{Deserialize, Serialize};
 
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ResultLaunchAttack {
-    pub launcher_id_name: String,
-    pub atk_name: String,
-    pub new_game_atk_effects: Vec<GameAtkEffect>,
-    pub is_crit: bool,
-    pub all_dodging: Vec<DodgeInfo>,
-    pub is_boss_atk: bool,
-    pub logs_end_of_round: Vec<LogData>,
-    pub logs_atk: Vec<LogData>,
-    /// Passive-triggered log entries for this turn (e.g. IsDamageTxHealNeedyAlly heal).
-    /// Included in `logs_atk` for the log sheet and also exposed here for the gameboard.
-    pub passive_logs: Vec<LogData>,
-    pub turn_nb: usize,
-    pub round_nb: usize,
-    /// True when the finishing blow was delivered by a damage-over-time effect (regen tick), not the direct attack.
-    pub is_dot_kill: bool,
-    /// Last attack name of the character killed by DOT (empty if not a DOT kill).
-    pub dying_char_last_atk: String,
-}
-
-/// The entry of the library.
-/// That object should be called to access to all the different functionalities.
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GameManager {
-    /// Game state of the current game
-    pub game_state: GameState,
-    /// Player manager
-    pub pm: PlayerManager,
-    /// Paths of the current game
-    pub game_paths: GamePaths,
-    /// logs of the game, to display in the log sheet
-    pub logs: Vec<LogData>,
-    /// Current scenario of the game, to adapt the behavior of the fight
-    pub current_scenario: Scenario,
-    /// all scenarios
-    pub all_scenarios: Vec<Scenario>,
-    /// State of the different scenarios, to know which scenario is available for the player and to adapt the behavior of the fight
-    pub states_scenarios: HashMap<String, ScenarioState>,
-    /// End of scenario
-    pub end_of_scenario: EndOfScenario,
-}
+use super::{GameManager, ResultLaunchAttack};
 
 impl GameManager {
-    /// Create a new game manager with the given path for the offline files and the default active characters
-    pub fn new<P: AsRef<Path>>(
-        path: P,
-        equipment_table: HashMap<EquipmentJsonKey, Vec<Equipment>>,
-        scenarios: Vec<Scenario>,
-    ) -> GameManager {
-        // if path is empty, use the default one
-        let mut new_path = path.as_ref();
-        if new_path.as_os_str().is_empty() {
-            new_path = &OFFLINE_ROOT;
-        }
-        // create game state
-        let game_state = GameState::new();
-        let game_name = game_state.game_name.clone();
-
-        // scenarios state
-        let mut states_scenarios = HashMap::new();
-        for scenario in &scenarios {
-            states_scenarios.insert(scenario.name.clone(), ScenarioState::NotStarted);
-        }
-
-        GameManager {
-            game_state,
-            pm: PlayerManager::new(equipment_table),
-            game_paths: GamePaths::new(new_path, &game_name),
-            logs: Vec::new(),
-            current_scenario: Scenario::default(),
-            all_scenarios: scenarios,
-            states_scenarios,
-            end_of_scenario: EndOfScenario::default(),
-        }
-    }
-
-    /// Set active bosses from the current scenario's boss patterns.
-    /// Bosses whose name matches a pattern in the current scenario are cloned and
-    /// pushed into `pm.active_bosses` with a unique id_name (`"<name>_#<n>"`).
-    pub fn set_active_bosses(&mut self, all_bosses: &[Character]) {
-        self.current_scenario
-            .boss_patterns
-            .iter()
-            .for_each(|(boss_name, _)| {
-                if let Some(b) = all_bosses.iter().find(|b| b.db_full_name == *boss_name) {
-                    let mut boss_to_push = b.clone();
-                    boss_to_push.id_name = format!(
-                        "{}_#{}",
-                        boss_to_push.db_full_name,
-                        1 + self
-                            .pm
-                            .get_nb_of_active_bosses_by_name(&boss_to_push.db_full_name)
-                    );
-                    self.pm.active_bosses.push(boss_to_push);
-                } else {
-                    tracing::warn!("Boss {} not found in data manager, skipping it", boss_name);
-                }
-            });
-    }
-
-    pub fn load_next_scenario(&mut self) -> Result<()> {
-        // update current scenario state
-        if let Some((_, state)) = self
-            .states_scenarios
-            .iter_mut()
-            .find(|(name, _)| *name == &self.current_scenario.name)
-        {
-            *state = ScenarioState::Completed;
-        }
-        let current_level = self.current_scenario.level;
-        let current_universe = self.current_scenario.universe.clone();
-        // get the next scenario with the next level in the same universe.
-        // When current_universe is empty (default scenario at game start) skip the
-        // universe filter so the first real scenario is found regardless of universe.
-        let Some(scenario) = self
-            .all_scenarios
-            .iter()
-            .find(|s| {
-                s.level == current_level + 1
-                    && (current_universe.is_empty() || s.universe == current_universe)
-            })
-            .cloned()
-        else {
-            return Err(anyhow::anyhow!(
-                "No next scenario found for level {}",
-                current_level + 1
-            ));
-        };
-        // update scenario state in map
-        if let Some((_, state)) = self
-            .states_scenarios
-            .iter_mut()
-            .find(|(name, _)| *name == &scenario.name)
-        {
-            *state = ScenarioState::InProgress;
-        }
-        // update current scenario
-        self.current_scenario = scenario;
-
-        if self.current_scenario.level > 1 {
-            // accumulate kills from the completed scenario before clearing
-            let scenario_kills = self
-                .pm
-                .active_bosses
-                .iter()
-                .filter(|b| b.stats.is_dead().unwrap_or(false))
-                .count();
-            self.game_state.accumulated_kills += scenario_kills;
-            // clear previous scenario
-            self.game_state.clear_scenario();
-            self.pm.clear_scenario();
-            // set active bosses for the new scenario from the stored roster
-            // do it before start new turn and after clearing a scenario
-            let all_bosses = self.pm.all_bosses.clone();
-            self.set_active_bosses(&all_bosses);
-            let _ = self.start_new_turn();
-        } else {
-            // set active bosses for the new scenario from the stored roster
-            let all_bosses = self.pm.all_bosses.clone();
-            self.set_active_bosses(&all_bosses);
-        }
-
-        Ok(())
-    }
-
-    pub fn all_scenarios_completed(&self) -> bool {
-        self.states_scenarios
-            .values()
-            .all(|state| *state == ScenarioState::Completed)
-    }
-
-    /// Start the game by starting a new turn
-    pub fn start_game(&mut self) {
-        // Start a new turn
-        let _ = self.start_new_turn();
-    }
-
-    /// Process the start of a new turn:
-    /// - Process the order of the players to play
-    /// - Increment the turn number
-    /// - Reset the round number
-    ///
-    /// Return a boolean to know if the new turn has been started and the logs of the new round if it is the case
-    pub fn start_new_turn(&mut self) -> (bool, Vec<LogData>) {
-        // For each turn now
-        // Process the order of the players
-        self.process_order_to_play();
-        self.game_state.start_new_turn();
-        self.pm.start_new_turn(self.game_state.current_turn_nb == 1);
-
-        self.new_round()
-    }
-
-    /// Process the order of the players to play by sorting them by speed and adding the supplementary atk turns for the heroes and the bosses
-    pub fn process_order_to_play(&mut self) {
-        // to be improved with stats
-        // one player can play several times as well in different order
-        self.game_state.order_to_play.clear();
-
-        // add heroes
-        // sort by speed descending (highest speed acts first)
-        self.pm
-            .active_heroes
-            .sort_by(|a, b| b.stats.all_stats[SPEED].cmp(&a.stats.all_stats[SPEED]));
-        let mut dead_heroes = Vec::new();
-        for hero in &self.pm.active_heroes {
-            if !hero.stats.is_dead().unwrap_or(false) {
-                self.game_state.order_to_play.push(hero.id_name.clone());
-            } else {
-                dead_heroes.push(hero.id_name.clone());
-            }
-        }
-        // add dead heroes
-        for name in dead_heroes {
-            self.game_state.order_to_play.push(name);
-        }
-        // add bosses
-        // sort by speed descending (highest speed acts first)
-        self.pm
-            .active_bosses
-            .sort_by(|a, b| b.stats.all_stats[SPEED].cmp(&a.stats.all_stats[SPEED]));
-        for boss in &self.pm.active_bosses {
-            if !boss.stats.is_dead().unwrap_or(false) {
-                self.game_state.order_to_play.push(boss.id_name.clone());
-            }
-        }
-        // supplementary atks to be added
-        let supp_rounds_heroes = self.pm.process_sup_atk_turn(CharacterKind::Hero);
-        let supp_rounds_bosses = self.pm.process_sup_atk_turn(CharacterKind::Boss);
-        self.game_state.order_to_play.extend(supp_rounds_heroes);
-        self.game_state.order_to_play.extend(supp_rounds_bosses);
-    }
-
-    pub fn new_round(&mut self) -> (bool, Vec<LogData>) {
-        self.game_state.new_round();
-        // Still round to play
-        if self.game_state.current_round > self.game_state.order_to_play.len() {
-            return (
-                false,
-                vec![LogData {
-                    message: "End of turn has been reached".to_string(),
-                    ..Default::default()
-                }],
-            );
-        }
-        let Ok(mut logs) = self.pm.update_current_player_on_new_round(
-            &self.game_state,
-            &self.game_state.order_to_play[self.game_state.current_round - 1],
-        ) else {
-            // return the error of update_current_player
-            return (
-                false,
-                vec![LogData {
-                    message: "Error while updating current player".to_string(),
-                    ..Default::default()
-                }],
-            );
-        };
-
-        if self.pm.current_player.stats.is_dead() == Some(true) {
-            let (all_heroes_dead, all_bosses_dead) = self.pm.check_end_of_game();
-            if all_heroes_dead {
-                self.game_state.status = GameStatus::EndOfGame;
-                return (false, logs);
-            } else if all_bosses_dead {
-                self.game_state.status = GameStatus::EndOfScenario;
-                self.process_end_of_scenario();
-                return (false, logs);
-            }
-            return self.new_round();
-        }
-
-        self.pm.reset_targeted_character();
-
-        // Insert a round-separator at the front so the log sheet can group events per round
-        logs.insert(
-            0,
-            LogData {
-                message: format!(
-                    "\u{1f501} Turn {} — Round {}",
-                    self.game_state.current_turn_nb, self.game_state.current_round
-                ),
-                color: crate::common::log_data::const_colors::LIGHT_BLUE.to_owned(),
-            },
-        );
-
-        (true, logs)
-    }
-
     /// Launch an attack from the current player
     /// If atk_name is None and it is an auto round (boss), a random atk will be chosen
     /// Otherwise, if atk_name is None, no atk will be launched
@@ -631,200 +331,6 @@ impl GameManager {
         result_attack
     }
 
-    /// Process end-of-scenario rewards for every hero:
-    /// - Add loot items matching the hero's class (equipment checked against the equipment database,
-    ///   consumables and currency added directly)
-    /// - Add experience gained from all defeated bosses and level up (with stat update) as needed
-    /// - Automatically use all consumables in inventory (potions restore HP)
-    ///   Process end of scenario struct to be sent to the frontend with the rewards and the level up info
-    pub fn process_end_of_scenario(&mut self) {
-        // Total exp: sum from all bosses
-        let total_exp: u64 = self
-            .pm
-            .active_bosses
-            .iter()
-            .map(|boss| build_experience(&boss.rank, boss.level))
-            .sum();
-
-        let loots = self.current_scenario.loots.clone();
-        let equipment_table_flat: Vec<Equipment> = self
-            .pm
-            .equipment_table
-            .values()
-            .flatten()
-            .cloned()
-            .collect();
-
-        // prepare end of scenario
-        self.end_of_scenario.scenario_level = self.current_scenario.level;
-        self.end_of_scenario.characters_levelup.clear();
-        self.pm.active_heroes.iter().for_each(|hero| {
-            self.end_of_scenario.characters_levelup.push(LevelUp {
-                character_id_name: hero.id_name.clone(),
-                new_level: hero.level,
-                old_level: hero.level,
-            });
-        });
-
-        for i in 0..self.pm.active_heroes.len() {
-            let hero_class = self.pm.active_heroes[i].class.clone();
-
-            // Add loot according to class
-            for loot in &loots {
-                let class_matches =
-                    loot.classes.contains(&hero_class) || loot.classes.contains(&Class::Standard);
-                if !class_matches {
-                    continue;
-                }
-                match &loot.kind {
-                    LootType::Equipment => {
-                        if let Some(equipment) = equipment_table_flat
-                            .iter()
-                            .find(|e| e.unique_name == loot.name)
-                            .cloned()
-                        {
-                            self.pm.active_heroes[i]
-                                .inventory
-                                .add_equipment(&equipment, false);
-                        } else {
-                            tracing::warn!(
-                                "Equipment '{}' not found in equipment database",
-                                loot.name
-                            );
-                        }
-                    }
-                    LootType::Consumable => {
-                        // Consumables go to the shared party bag (handled below).
-                    }
-                    LootType::Currency => {
-                        self.pm.active_heroes[i].inventory.money += loot.level as u64;
-                    }
-                    LootType::Material => {}
-                }
-            }
-
-            // Add experience and level up if needed
-            self.pm.active_heroes[i].character_rounds_info.exp += total_exp;
-            while self.pm.active_heroes[i].character_rounds_info.exp
-                >= self.pm.active_heroes[i]
-                    .character_rounds_info
-                    .exp_to_next_level
-                && self.pm.active_heroes[i].level < ULTIMATE_LEVEL
-            {
-                self.pm.active_heroes[i].character_rounds_info.exp -= self.pm.active_heroes[i]
-                    .character_rounds_info
-                    .exp_to_next_level;
-                self.pm.active_heroes[i].level += 1;
-                self.pm.active_heroes[i].stats.update_stats_to_next_level();
-                // Recompute the threshold for the new level
-                self.pm.active_heroes[i]
-                    .character_rounds_info
-                    .exp_to_next_level = build_exp_to_next_level(
-                    &self.pm.active_heroes[i].rank,
-                    &self.pm.active_heroes[i].class,
-                    self.pm.active_heroes[i].level,
-                );
-                // update end of scenario
-                if let Some(level_up) = self
-                    .end_of_scenario
-                    .characters_levelup
-                    .iter_mut()
-                    .find(|lu| lu.character_id_name == self.pm.active_heroes[i].id_name)
-                {
-                    level_up.new_level = self.pm.active_heroes[i].level;
-                }
-            }
-        }
-
-        // Add consumable loot to the shared party bag (once per loot item, not per hero).
-        for loot in &loots {
-            if loot.kind != LootType::Consumable {
-                continue;
-            }
-            let any_hero_matches = self.pm.active_heroes.iter().any(|hero| {
-                loot.classes.contains(&hero.class) || loot.classes.contains(&Class::Standard)
-            });
-            if any_hero_matches && let Some(consumable) = build_consumable_by_name(&loot.name) {
-                self.pm.party_consumables.push(consumable);
-            }
-        }
-    }
-
-    fn process_no_atk_launched(&mut self) -> ResultLaunchAttack {
-        // Capture launcher identity before eval_end_of_round() may advance current_player.
-        let launcher_id_name = self.pm.current_player.id_name.clone();
-        let is_boss_atk = self.pm.current_player.is_boss_atk();
-        let turn_nb = self.game_state.current_turn_nb;
-        let round_nb = self.game_state.current_round;
-
-        self.pm
-            .current_player
-            .character_rounds_info
-            .actions_done_in_round += 1;
-        let logs_atk = vec![LogData {
-            message: "No attack launched".to_string(),
-            color: DARK_RED.to_string(),
-        }];
-        let logs_end_of_round = self.eval_end_of_round(logs_atk.clone());
-        let result = ResultLaunchAttack {
-            launcher_id_name,
-            is_boss_atk,
-            logs_end_of_round,
-            logs_atk,
-            turn_nb,
-            round_nb,
-            ..Default::default()
-        };
-        // Mirror what launch_attack() does for real attacks so ra.atk_name is
-        // empty and the gameboard consumable-action banner branch is reachable.
-        self.game_state.last_result_atk = result.clone();
-        result
-    }
-
-    /// Evaluate the end of the round by checking if the game is finished,
-    ///  if a new round should start or if a new turn should start,
-    ///  and return the logs to display for the new round if it is the case
-    fn eval_end_of_round(&mut self, logs_atk: Vec<LogData>) -> Vec<LogData> {
-        let mut output_logs = vec![];
-        let (all_heroes_dead, all_bosses_dead) = self.pm.check_end_of_game();
-        if all_heroes_dead {
-            self.game_state.status = GameStatus::EndOfGame;
-        } else if all_bosses_dead {
-            self.game_state.status = GameStatus::EndOfScenario;
-            self.process_end_of_scenario();
-        } else {
-            let (is_new_round, logs) = self.new_round();
-            output_logs.extend(logs);
-            // new_round may have triggered EndOfScenario/EndOfGame (e.g. boss killed by DOT)
-            if matches!(
-                self.game_state.status,
-                GameStatus::EndOfScenario | GameStatus::EndOfGame
-            ) {
-                // Status already set inside new_round; nothing more to do
-            } else if is_new_round {
-                self.game_state.status = GameStatus::StartRound;
-            } else {
-                let (is_new_turn, logs) = self.start_new_turn();
-                output_logs.extend(logs);
-                if matches!(
-                    self.game_state.status,
-                    GameStatus::EndOfScenario | GameStatus::EndOfGame
-                ) {
-                    // Status set inside start_new_turn via new_round
-                } else if is_new_turn {
-                    self.game_state.status = GameStatus::StartRound;
-                } else {
-                    self.game_state.status = GameStatus::EndOfGame;
-                }
-            }
-        }
-
-        self.logs.extend(output_logs.clone());
-        self.logs.extend(logs_atk.clone());
-
-        output_logs
-    }
-
     pub fn build_logs_atk(
         &self,
         all_dodging: &Vec<DodgeInfo>,
@@ -898,56 +404,13 @@ impl GameManager {
         }
         logs
     }
-
-    /// Check if it is the turn to a boss to play
-    /// HMI function
-    pub fn is_round_auto(&self) -> bool {
-        if self.game_state.current_round as i64 > 0
-            && self.game_state.current_round as i64 - 1 < self.game_state.order_to_play.len() as i64
-        {
-            let name = self.game_state.order_to_play[self.game_state.current_round - 1].clone();
-            if let Some(c) = self.pm.get_active_character(&name) {
-                return c.kind == CharacterKind::Boss;
-            }
-        }
-
-        false
-    }
-
-    /// Process the number of bosses that are attacking in a row in the current round, to know if it is the case to add a log for the new round with the info of the boss attack
-    /// boss should not be dead to be counted
-    /// used by dx-rpg
-    pub fn process_nb_bosses_atk_in_a_row(&self) -> i64 {
-        let mut count = 0;
-
-        if self.game_state.current_round as i64 > 0
-            && self.game_state.current_round as i64 - 1 < self.game_state.order_to_play.len() as i64
-        {
-            // Start from current_round and go to the end
-            for i in self.game_state.current_round - 1..self.game_state.order_to_play.len() {
-                let name = &self.game_state.order_to_play[i];
-
-                if let Some(c) = self.pm.get_active_character(name) {
-                    if c.kind == CharacterKind::Boss && c.stats.is_dead() != Some(true) {
-                        count += 1;
-                    } else {
-                        break; // Stop counting when a non-Boss is found
-                    }
-                } else {
-                    break; // Stop counting if character doesn't exist
-                }
-            }
-        }
-
-        count
-    }
 }
 
 #[cfg(test)]
 mod tests {
     use crate::character_mod::attack_type::AttackType;
     use crate::character_mod::buffers::{BufKinds, Buffer};
-    use crate::character_mod::character::CharacterKind;
+
     use crate::character_mod::class::Class;
     use crate::character_mod::rank::Rank;
     use crate::common::constants::attak_const::COEFF_CRIT_DMG;
@@ -955,126 +418,8 @@ mod tests {
     use crate::common::log_data::const_colors::DARK_RED;
     use crate::server::game_manager::LogData;
     use crate::server::game_state::GameStatus;
-    use crate::testing::testing_all_characters::{
-        self, testing_game_manager, testing_test_ally1_vs_test_boss1,
-    };
-    use crate::{
-        common::constants::{character_const::SPEED_THRESHOLD, stats_const::*},
-        testing::testing_atk::*,
-    };
-
-    #[test]
-    fn unit_process_order_to_play() {
-        let mut gm = testing_game_manager();
-        let old_speed = gm
-            .pm
-            .get_mut_active_hero_character("test_#1")
-            .cloned()
-            .unwrap()
-            .stats
-            .all_stats[SPEED]
-            .clone();
-        gm.process_order_to_play();
-        let new_speed = gm
-            .pm
-            .get_mut_active_hero_character("test_#1")
-            .cloned()
-            .unwrap()
-            .stats
-            .all_stats[SPEED]
-            .clone();
-        // only one supplementary attack per turn: test2_#1 (fastest hero, speed 312) qualifies;
-        // test_#1 (212) is skipped because process_sup_atk_turn returns after the first hit.
-        assert_eq!(gm.game_state.order_to_play.len(), 5);
-        // descending speed sort: test2_#1 (312) before test_#1 (212)
-        assert_eq!(gm.game_state.order_to_play[0], "test2_#1");
-        assert_eq!(gm.game_state.order_to_play[1], "test_#1");
-        // descending speed sort: boss2 (15) before boss1 (11)
-        assert_eq!(gm.game_state.order_to_play[2], "test_boss2_#1");
-        assert_eq!(gm.game_state.order_to_play[3], "test_boss1_#1");
-        // only test2_#1 gets the supplementary slot
-        assert_eq!(gm.game_state.order_to_play[4], "test2_#1");
-        // test_#1 speed is unchanged (it did NOT get the supplementary slot)
-        assert_eq!(old_speed.current, new_speed.current);
-        assert_eq!(old_speed.max, new_speed.max);
-        assert_eq!(old_speed.max_raw, new_speed.max_raw);
-        assert_eq!(old_speed.current_raw, new_speed.current_raw);
-        // test2_#1 had its speed reset (312 - SPEED_THRESHOLD = 212)
-        let new_test2_speed = gm
-            .pm
-            .get_mut_active_hero_character("test2_#1")
-            .cloned()
-            .unwrap()
-            .stats
-            .all_stats[SPEED]
-            .clone();
-        assert_eq!(312 - SPEED_THRESHOLD, new_test2_speed.current);
-        // one hero player is dead — use name-based kill so the index stays stable after sort
-        gm.pm
-            .get_mut_active_hero_character("test_#1")
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current = 0;
-        gm.process_order_to_play();
-        assert_eq!(gm.game_state.order_to_play.len(), 5);
-        assert_eq!(gm.game_state.order_to_play[0], "test2_#1");
-        assert_eq!(gm.game_state.order_to_play[1], "test_#1");
-        assert_eq!(gm.game_state.order_to_play[2], "test_boss2_#1");
-        assert_eq!(gm.game_state.order_to_play[3], "test_boss1_#1");
-        assert_eq!(gm.game_state.order_to_play[4], "test2_#1");
-        // boss is dead — use name-based kill; descending sort puts boss2 at index 0
-        gm.pm
-            .get_mut_active_boss_character("test_boss1_#1")
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current = 0;
-        gm.process_order_to_play();
-        assert_eq!(gm.game_state.order_to_play.len(), 4);
-        assert_eq!(gm.game_state.order_to_play[0], "test2_#1");
-        assert_eq!(gm.game_state.order_to_play[1], "test_#1");
-        assert_eq!(gm.game_state.order_to_play[2], "test_boss2_#1");
-        assert_eq!(gm.game_state.order_to_play[3], "test2_#1");
-    }
-
-    #[test]
-    fn unit_add_sup_atk_turn() {
-        let mut gm = testing_all_characters::testing_game_manager();
-        let hero = gm.pm.active_heroes.first_mut().unwrap();
-        hero.stats.all_stats.get_mut(SPEED).unwrap().current = 300;
-        let boss = gm.pm.active_bosses.first_mut().unwrap();
-        boss.stats.all_stats.get_mut(SPEED).unwrap().current = 10;
-        let result = gm.pm.process_sup_atk_turn(CharacterKind::Hero);
-        // only one supplementary attack per call — the first qualifying hero
-        assert_eq!(result.len(), 1);
-    }
-
-    #[test]
-    fn unit_new_round() {
-        let mut gm = testing_all_characters::testing_game_manager();
-        let result = gm.start_new_turn();
-        assert!(result.0);
-        assert_eq!(gm.game_state.current_round, 1);
-
-        // test current player -test- is dead - round for boss is starting
-        gm.game_state.current_round = 0;
-        gm.pm.active_heroes[0].stats.all_stats[HP].current = 0;
-        let result = gm.new_round();
-        assert!(result.0);
-        assert_eq!(gm.game_state.current_round, 2);
-        // test current round > table order to play
-        gm.game_state.current_round = 1000;
-        let result = gm.new_round();
-        assert!(!result.0);
-        // character name in orderToplay list is not a player
-        gm.game_state.order_to_play.clear();
-        gm.game_state.order_to_play.push("unknown".to_owned());
-        gm.game_state.current_round = 0;
-        let result = gm.new_round();
-        assert!(!result.0);
-    }
-
+    use crate::testing::testing_all_characters::{self, testing_test_ally1_vs_test_boss1};
+    use crate::{common::constants::stats_const::*, testing::testing_atk::*};
     #[test]
     fn unit_launch_attack_none_atk_hero() {
         let (mut gm, _hero_launcher_id_name, _target_id_name) = testing_test_ally1_vs_test_boss1();
@@ -2035,583 +1380,6 @@ mod tests {
         }
     }
 
-    #[test]
-    fn unit_load_next_scenario() {
-        use crate::server::scenario::ScenarioState;
-
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-
-        // dxrpg loads lotr scenarios; states start as NotStarted
-        let stage1_name = "Patrouille Gobeline".to_owned();
-        let stage2_name = "Embuscade Gobeline".to_owned();
-        assert_eq!(gm.states_scenarios[&stage1_name], ScenarioState::NotStarted);
-        assert_eq!(gm.states_scenarios[&stage2_name], ScenarioState::NotStarted);
-
-        // set stage 1 as current (simulates game start on stage 1)
-        let stage1 = gm
-            .all_scenarios
-            .iter()
-            .find(|s| s.name == stage1_name)
-            .cloned()
-            .unwrap();
-        gm.current_scenario = stage1;
-        gm.states_scenarios
-            .insert(stage1_name.clone(), ScenarioState::InProgress);
-
-        // damage heroes and drain their energy to verify restoration on next scenario
-        for hero in gm.pm.active_heroes.iter_mut() {
-            hero.stats.get_mut_value(HP).current = 1;
-            hero.stats.get_mut_value(MANA).current = 0;
-            hero.stats.get_mut_value(VIGOR).current = 0;
-            hero.stats.get_mut_value(BERSERK).current = 0;
-        }
-
-        // load stage 2
-        let result = gm.load_next_scenario();
-        assert!(result.is_ok(), "loading stage 2 should succeed");
-
-        // stage 1 must be Completed
-        assert_eq!(
-            gm.states_scenarios[&stage1_name],
-            ScenarioState::Completed,
-            "stage 1 should be Completed after loading stage 2"
-        );
-        // stage 2 must be InProgress
-        assert_eq!(
-            gm.states_scenarios[&stage2_name],
-            ScenarioState::InProgress,
-            "stage 2 should be InProgress after being loaded"
-        );
-        // current_scenario must be stage 2
-        assert_eq!(gm.current_scenario.name, stage2_name);
-
-        // active_bosses count must equal the stage 2 boss patterns
-        assert_eq!(
-            gm.pm.active_bosses.len(),
-            gm.current_scenario.boss_patterns.len(),
-            "active_bosses should match stage 2 boss patterns count"
-        );
-
-        // heroes must have HP, energy and no effects restored to max
-        for hero in gm.pm.active_heroes.iter() {
-            assert_eq!(
-                hero.stats.all_stats[HP].current, hero.stats.all_stats[HP].max,
-                "hero {} HP should be restored to max",
-                hero.db_full_name
-            );
-            assert_eq!(
-                hero.stats.all_stats[MANA].current, hero.stats.all_stats[MANA].max,
-                "hero {} Mana should be restored to max",
-                hero.db_full_name
-            );
-            assert_eq!(
-                hero.stats.all_stats[VIGOR].current, hero.stats.all_stats[VIGOR].max,
-                "hero {} Vigor should be restored to max",
-                hero.db_full_name
-            );
-            assert_eq!(
-                hero.stats.all_stats[BERSERK].current, 0,
-                "hero {} Berserk should NOT be restored on scenario load",
-                hero.db_full_name
-            );
-            assert!(
-                hero.character_rounds_info.all_effects.is_empty(),
-                "hero {} should have no active effects after scenario transition",
-                hero.db_full_name
-            );
-        }
-
-        // all_scenarios_completed returns false (stage 2 still in progress)
-        assert!(!gm.all_scenarios_completed());
-    }
-
-    #[test]
-    fn unit_set_active_bosses() {
-        use crate::testing::testing_all_characters::dxrpg_dm;
-
-        let dm = dxrpg_dm();
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-
-        // set stage 1 as current scenario so boss_patterns are in scope
-        let stage1 = gm
-            .all_scenarios
-            .iter()
-            .find(|s| s.level == 1)
-            .cloned()
-            .unwrap();
-        gm.current_scenario = stage1;
-
-        // no bosses yet
-        gm.pm.active_bosses.clear();
-        assert_eq!(gm.pm.active_bosses.len(), 0);
-
-        gm.set_active_bosses(&dm.all_bosses);
-
-        // the number of active bosses must match the number of boss_patterns entries
-        // that have a matching entry in dm.all_bosses
-        let expected = gm
-            .current_scenario
-            .boss_patterns
-            .keys()
-            .filter(|name| dm.all_bosses.iter().any(|b| &b.db_full_name == *name))
-            .count();
-        assert_eq!(
-            gm.pm.active_bosses.len(),
-            expected,
-            "active_bosses count should match boss_patterns with a known boss"
-        );
-
-        // each active boss must have the correct id_name suffix format
-        for boss in &gm.pm.active_bosses {
-            assert!(
-                boss.id_name.contains("_#"),
-                "id_name '{}' should contain '_#'",
-                boss.id_name
-            );
-        }
-    }
-
-    // -------------------------------------------------------------------------
-    // process_end_of_scenario tests
-    // -------------------------------------------------------------------------
-
-    #[test]
-    fn unit_end_of_scenario_equipment_loot_matching_class() {
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-        // Both test heroes are Standard class.
-        // Create a scenario with one equipment loot targeting Standard heroes.
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            level: 1,
-            loots: vec![Loot {
-                name: "starting right weapon".to_string(),
-                kind: LootType::Equipment,
-                rank: Rank::Common,
-                level: 1,
-                classes: vec![Class::Standard],
-            }],
-            universe: String::new(),
-        };
-
-        gm.process_end_of_scenario();
-
-        // Both heroes must now have the equipment in their inventory
-        for hero in &gm.pm.active_heroes {
-            let has_it = hero
-                .inventory
-                .equipments
-                .values()
-                .flatten()
-                .any(|e| e.unique_name == "starting right weapon");
-            assert!(
-                has_it,
-                "hero '{}' should have received the equipment",
-                hero.id_name
-            );
-        }
-    }
-
-    #[test]
-    fn unit_end_of_scenario_equipment_loot_no_class_match() {
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-        // Both test heroes are Standard.
-        // Equipment loot only for Warrior → heroes must NOT receive an extra copy.
-        let belts_before: Vec<usize> = gm
-            .pm
-            .active_heroes
-            .iter()
-            .map(|h| {
-                h.inventory
-                    .equipments
-                    .values()
-                    .flatten()
-                    .filter(|e| e.unique_name == "starting belt")
-                    .count()
-            })
-            .collect();
-
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            level: 1,
-            loots: vec![Loot {
-                name: "starting belt".to_string(),
-                kind: LootType::Equipment,
-                rank: Rank::Common,
-                level: 1,
-                classes: vec![Class::Warrior],
-            }],
-            universe: String::new(),
-        };
-
-        gm.process_end_of_scenario();
-
-        for (idx, hero) in gm.pm.active_heroes.iter().enumerate() {
-            let belts_after = hero
-                .inventory
-                .equipments
-                .values()
-                .flatten()
-                .filter(|e| e.unique_name == "starting belt")
-                .count();
-            assert_eq!(
-                belts_after, belts_before[idx],
-                "hero '{}' belt count should not change (class mismatch)",
-                hero.id_name
-            );
-        }
-    }
-
-    #[test]
-    fn unit_end_of_scenario_equipment_loot_unknown_equipment() {
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-        // Record initial equipment count per hero
-        let equip_before: Vec<usize> = gm
-            .pm
-            .active_heroes
-            .iter()
-            .map(|h| h.inventory.equipments.values().map(|v| v.len()).sum())
-            .collect();
-
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            level: 1,
-            loots: vec![Loot {
-                name: "non_existent_equipment".to_string(),
-                kind: LootType::Equipment,
-                rank: Rank::Common,
-                level: 1,
-                classes: vec![Class::Standard],
-            }],
-            universe: String::new(),
-        };
-
-        // Must not panic; unknown equipment is just warned about and skipped
-        gm.process_end_of_scenario();
-
-        for (idx, hero) in gm.pm.active_heroes.iter().enumerate() {
-            let total_equip: usize = hero.inventory.equipments.values().map(|v| v.len()).sum();
-            assert_eq!(
-                total_equip, equip_before[idx],
-                "hero '{}' equipment count should not change for unknown loot name",
-                hero.id_name
-            );
-        }
-    }
-
-    #[test]
-    fn unit_end_of_scenario_consumable_loot() {
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            level: 1,
-            loots: vec![Loot {
-                name: "potion".to_string(),
-                kind: LootType::Consumable,
-                rank: Rank::Common,
-                level: 1,
-                classes: vec![Class::Standard],
-            }],
-            universe: String::new(),
-        };
-
-        gm.process_end_of_scenario();
-
-        // Consumables must go to the shared party bag — not to individual heroes.
-        let in_party_bag = gm.pm.party_consumables.iter().any(|c| c.name == "potion");
-        assert!(
-            in_party_bag,
-            "consumable should land in the party bag, not in individual inventories"
-        );
-
-        // Personal inventories must be untouched.
-        for hero in &gm.pm.active_heroes {
-            let in_personal_bag = hero
-                .inventory
-                .consumables
-                .iter()
-                .any(|c| c.name == "potion");
-            assert!(
-                !in_personal_bag,
-                "hero '{}' should NOT have the consumable in their personal bag",
-                hero.id_name
-            );
-        }
-    }
-
-    #[test]
-    fn unit_loot_consumables_use_shop_definitions() {
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-
-        for (potion_name, rank) in [
-            ("potion of resurrection", Rank::Advanced),
-            ("mana potion", Rank::Intermediate),
-            ("vigor potion", Rank::Common),
-            ("berserk potion", Rank::Advanced),
-        ] {
-            let original_len = gm.pm.party_consumables.len();
-            gm.current_scenario = Scenario {
-                name: "test".to_string(),
-                description: "test".to_string(),
-                boss_patterns: HashMap::new(),
-                level: 1,
-                loots: vec![Loot {
-                    name: potion_name.to_string(),
-                    kind: LootType::Consumable,
-                    rank,
-                    level: 1,
-                    classes: vec![Class::Standard],
-                }],
-                universe: String::new(),
-            };
-            gm.process_end_of_scenario();
-            let found = gm
-                .pm
-                .party_consumables
-                .iter()
-                .skip(original_len)
-                .any(|c| c.name == potion_name);
-            assert!(
-                found,
-                "'{potion_name}' should be in party bag after end_of_scenario"
-            );
-        }
-    }
-
-    #[test]
-    fn unit_end_of_scenario_currency_loot() {
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            level: 1,
-            loots: vec![Loot {
-                name: "gold".to_string(),
-                kind: LootType::Currency,
-                rank: Rank::Common,
-                level: 100,
-                classes: vec![Class::Standard],
-            }],
-            universe: String::new(),
-        };
-
-        // Test heroes already have money: 100 in their JSON
-        let money_before: Vec<u64> = gm
-            .pm
-            .active_heroes
-            .iter()
-            .map(|h| h.inventory.money)
-            .collect();
-
-        gm.process_end_of_scenario();
-
-        for (idx, hero) in gm.pm.active_heroes.iter().enumerate() {
-            assert_eq!(
-                hero.inventory.money,
-                money_before[idx] + 100,
-                "hero '{}' should have received 100 gold",
-                hero.id_name
-            );
-        }
-    }
-
-    #[test]
-    fn unit_end_of_scenario_exp_and_level_up() {
-        // Test setup: 2 bosses, each rank Common level 1 → 100 exp each → 200 total
-        //
-        // "test" hero:  exp=50, exp_to_next_level(Common, Standard, 1)=100
-        //   50 + 200 = 250 → level-up to 2 (exp=150), new threshold=200 → 150 < 200 → stop at level 2
-        //
-        // "test2" hero: exp=0,  exp_to_next_level(Common, Standard, 1)=100
-        //   0 + 200 = 200 → level-up to 2 (exp=100), new threshold=200 → 100 < 200 → stop at level 2
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            loots: vec![],
-            level: 1,
-            universe: String::new(),
-        };
-
-        let old_hp_max: Vec<u64> = gm
-            .pm
-            .active_heroes
-            .iter()
-            .map(|h| h.stats.all_stats[HP].max)
-            .collect();
-
-        gm.process_end_of_scenario();
-
-        for (idx, hero) in gm.pm.active_heroes.iter().enumerate() {
-            assert_eq!(
-                hero.level, 2,
-                "hero '{}' should be level 2 after 200 exp (dynamic threshold grows to 200 at level 2)",
-                hero.id_name
-            );
-            // exp_to_next_level must now reflect the new level
-            assert_eq!(
-                hero.character_rounds_info.exp_to_next_level,
-                200, // build_exp_to_next_level(Common, Standard, 2) = 200
-                "hero '{}' exp_to_next_level should be 200 at level 2",
-                hero.id_name
-            );
-            // Stats must have been updated upward on level-up
-            assert!(
-                hero.stats.all_stats[HP].max > old_hp_max[idx],
-                "hero '{}' HP max should have increased after leveling up",
-                hero.id_name
-            );
-        }
-        // assess end of scenario LevelUp
-        assert_eq!(gm.end_of_scenario.characters_levelup.len(), 2); // 2 heroes
-        gm.end_of_scenario.characters_levelup.iter().for_each(|lu| {
-            assert_eq!(
-                lu.new_level, 2,
-                "LevelUp record should show new level 2 for hero '{}'",
-                lu.character_id_name
-            );
-            assert_eq!(
-                lu.old_level, 1,
-                "LevelUp record should show old level 1 for hero '{}'",
-                lu.character_id_name
-            );
-        });
-    }
-
-    #[test]
-    fn unit_end_of_scenario_no_level_up_when_exp_insufficient() {
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let mut gm = testing_game_manager();
-        // Remove all bosses so total_exp = 0
-        gm.pm.active_bosses.clear();
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            loots: vec![],
-            level: 1,
-            universe: String::new(),
-        };
-
-        let levels_before: Vec<u64> = gm.pm.active_heroes.iter().map(|h| h.level).collect();
-
-        gm.process_end_of_scenario();
-
-        for (idx, hero) in gm.pm.active_heroes.iter().enumerate() {
-            assert_eq!(
-                hero.level, levels_before[idx],
-                "hero '{}' should not have leveled up with 0 exp",
-                hero.id_name
-            );
-        }
-    }
-
-    #[test]
-    fn unit_end_of_scenario_triggered_via_game_flow() {
-        // Verify that eval_end_of_round sets EndOfScenario and processes rewards
-        // when all bosses are killed in a single hit.
-        use crate::character_mod::class::Class;
-        use crate::character_mod::loot::{Loot, LootType};
-        use crate::character_mod::rank::Rank;
-        use crate::server::scenario::Scenario;
-        use std::collections::HashMap;
-
-        let (mut gm, _, _) = testing_test_ally1_vs_test_boss1();
-
-        gm.current_scenario = Scenario {
-            name: "test".to_string(),
-            description: "test".to_string(),
-            boss_patterns: HashMap::new(),
-            level: 1,
-            loots: vec![Loot {
-                name: "gold".to_string(),
-                kind: LootType::Currency,
-                rank: Rank::Common,
-                level: 50,
-                classes: vec![Class::Standard],
-            }],
-            universe: String::new(),
-        };
-
-        // Kill all bosses
-        for boss in gm.pm.active_bosses.iter_mut() {
-            boss.stats.all_stats.get_mut(HP).unwrap().current = 0;
-        }
-
-        // Set target and launch — eval_end_of_round sees all bosses dead
-        gm.pm
-            .get_mut_active_boss_character("test_boss1_#1")
-            .unwrap()
-            .character_rounds_info
-            .is_current_target = true;
-        gm.launch_attack(None);
-
-        assert_eq!(
-            gm.game_state.status,
-            GameStatus::EndOfScenario,
-            "status should be EndOfScenario"
-        );
-        // Rewards must have been processed: each Standard hero got 50 gold on top of their starting 100
-        for hero in &gm.pm.active_heroes {
-            assert!(
-                hero.inventory.money >= 50,
-                "hero '{}' should have received 50 gold after end-of-scenario",
-                hero.id_name
-            );
-        }
-    }
-
     /// Fracas Marteau deals self-damage via its buffer effects on the caster.
     /// This test verifies that a low-HP hero is killed by the self-damage component.
     #[test]
@@ -3038,28 +1806,6 @@ mod tests {
     }
 
     // ── Rameau Guérisseur tests ────────────────────────────────────────────────
-    // The attack applies:
-    //   1. DecreasingRateOnTurn HP HOT on an individual ally (nb_turns=4, value=3)
-    //   2. ChangeMaxStat +10% Magic power on the same target (nb_turns=4)
-    //
-    // Test char stats: Magical Power max=30 (20 raw + 10 from starting_gloves equipment),
-    //   HP max=135, Mana 200/200.
-    // HOT per-tick = applies * (buffer.value + magic_power_current / nb_turns)
-    //              = applies * (3 + 30/4) = applies * 10  (integer division)
-    // applies comes from process_decrease_on_turn(value=3): always at least 1
-    // (first threshold is 100%), at most 3.  Per-tick ∈ [10, 30].
-    //
-    // The HOT then fires probabilistically on subsequent turns:
-    //   T2 counter=1: threshold = (3−1+1)/3 = 100 % (always)
-    //   T3 counter=2: threshold = (3−2+1)/3 =  67 %
-    //   T4 counter=3: threshold = (3−3+1)/3 =  33 %
-    // So the HOT fires 1–3 ticks after launch (not always 3).
-    //
-    // The DecreasingRateOnTurn effect also stores its applies count in ApplyEffectInit,
-    // which is then picked up by ALL subsequent effects in the same attack.  So
-    // the ChangeMaxStat full_amount = applies * 10, giving a magic
-    // power increase of 30 * (applies*10) / 100 = applies * 3.
-    // New magic power max ∈ [33, 36, 39] for applies ∈ [1, 2, 3].
 
     #[test]
     fn unit_rameau_guerisseur_initial_heal_range() {
@@ -3324,28 +2070,6 @@ mod tests {
         assert!(
             hot_ticks <= 3,
             "HOT must fire at most 3 times (T2–T4): fired {hot_ticks} times"
-        );
-    }
-
-    #[test]
-    fn unit_new_round_all_heroes_dead_end_of_game() {
-        let mut gm = testing_game_manager();
-        gm.start_game();
-        // Kill ALL heroes
-        for hero in &mut gm.pm.active_heroes {
-            hero.stats.all_stats[HP].current = 0;
-        }
-        // Make round 1 point to the first hero (who is dead)
-        gm.game_state.current_round = 0;
-        let (is_new_round, _logs) = gm.new_round();
-        assert!(
-            !is_new_round,
-            "dead player → should not start a new round normally"
-        );
-        assert_eq!(
-            gm.game_state.status,
-            GameStatus::EndOfGame,
-            "all heroes dead → EndOfGame"
         );
     }
